Fix help ticket close buttons timing out and allow ticket participants to close
Close buttons stopped responding after bot restarts because per-ticket views
were only re-registered from a startup API fetch. Replace them with
DynamicItem buttons matched by custom_id pattern, and allow the ticket
opener and thread participants (tribe owner, mentioned team) to close in
addition to moderators.
